requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
werkzeug==2.0.1
wtforms==2.3.3
bcrypt==3.2.0
python-magic==0.4.24